% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPTH2.R
\name{DEPTH2}
\alias{DEPTH2}
\title{## DEPTH2: An mRNA-based algorithm to evaluate intratumor heterogeneity (ITH) without reference to normal controls}
\usage{
DEPTH2(input_data)
}
\arguments{
\item{input_data}{A dataframe or matrix containing gene expression profiles in tumors, in which the row name is gene symbol or Entrez ID and the column name is tumor sample ID.}
}
\value{
A dataframe with 2 columns:
\item{Sample}{Tumor samples to be calculated.}
\item{DEPTH2_score}{The DEPTH2 score of each sample.}
}
\description{
\code{DEPTH2} DEPTH2 evaluates the ITH level of each bulk tumor sample based on gene expression profiles.
}
\details{
This function calculates the ITH score of a bulk tumor based on the standard deviations of absolute z-scored expression values of a set of genes in the tumor.
}
\author{
Dandan Song <dandan.song@stu.cpu.edu.cn> and Xiaosheng Wang <xiaosheng.wang@cpu.edu.cn>
}
